+++ trunk/cpsycle/audio/src/buffercontext.h
@@ -11,17 +11,21 @@ typedef struct {
 	psy_audio_Buffer* input;
 	psy_audio_Buffer* output;
 	psy_List* events;
-	unsigned int numsamples;
-	unsigned int numtracks;
+	uintptr_t numsamples;
+	uintptr_t numtracks;
 	struct psy_dsp_RMSVol* rmsvol;
 } psy_audio_BufferContext;
 
-void buffercontext_init(psy_audio_BufferContext*, psy_List* events, psy_audio_Buffer* input,
-	psy_audio_Buffer* output, unsigned int numsamples, unsigned int numtracks,
+void buffercontext_init(psy_audio_BufferContext*,
+	psy_List* events,
+	psy_audio_Buffer* input,
+	psy_audio_Buffer* output,
+	uintptr_t numsamples,
+	uintptr_t numtracks,
 	struct psy_dsp_RMSVol* rmsvol);
 void buffercontext_dispose(psy_audio_BufferContext*);
-unsigned int buffercontext_numsamples(psy_audio_BufferContext*);
-unsigned int buffercontext_numtracks(psy_audio_BufferContext*);
+uintptr_t buffercontext_numsamples(psy_audio_BufferContext*);
+uintptr_t buffercontext_numtracks(psy_audio_BufferContext*);
 float buffercontext_rmsvolume(psy_audio_BufferContext*);
